+++ src/com/liquidbol/commons/model/BillPayment.java
@@ -4,18 +4,23 @@
  * and open the template in the editor.
  */
 
-package com.liquidbol.model.commons;
+package com.liquidbol.commons.model;
 
+import com.liquidbol.db.persistence.EmployeeCrud;
+import com.liquidbol.db.persistence.PersistenceException;
 import java.io.Serializable;
 import java.sql.Date;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
- * Class that represents a service sale.
+ * Class that represents a bill payment.
  * @author Allan Leon
  */
-public class ServiceSale implements Serializable {
+public class BillPayment implements Serializable {
     
     private int id;
+    private Employee employee;
     private Date payDate;
     private Double amountPaid;
     private String obs;
@@ -23,12 +28,14 @@ public class ServiceSale implements Serializable {
     /**
      * Constructor method.
      * @param id
+     * @param employee
      * @param payDate
      * @param amountPaid
      * @param obs 
      */
-    public ServiceSale(int id, Date payDate, Double amountPaid, String obs) {
+    public BillPayment(int id, Employee employee, Date payDate, Double amountPaid, String obs) {
         this.id = id;
+        this.employee = employee;
         this.payDate = payDate;
         this.amountPaid = amountPaid;
         this.obs = obs;
@@ -39,6 +46,13 @@ public class ServiceSale implements Serializable {
      */
     public int getId() {
         return id;
+    }
+
+    /**
+     * @return the employee
+     */
+    public Employee getEmployee() {
+        return employee;
     }
 
     /**
@@ -70,6 +84,13 @@ public class ServiceSale implements Serializable {
     }
 
     /**
+     * @param employee the employee to set
+     */
+    public void setEmployee(Employee employee) {
+        this.employee = employee;
+    }
+
+    /**
      * @param payDate the payDate to set
      */
     public void setPayDate(Date payDate) {
@@ -89,11 +110,21 @@ public class ServiceSale implements Serializable {
     public void setObs(String obs) {
         this.obs = obs;
     }
+    
+    public void refresh() {
+        try {
+            employee = new EmployeeCrud().refresh(employee);
+        } catch (PersistenceException ex) {
+            Logger.getLogger(BillPayment.class.getName()).log(Level.SEVERE, null, ex);
+        } catch (ClassNotFoundException ex) {
+            Logger.getLogger(BillPayment.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
 
     @Override
     public int hashCode() {
         int hash = 3;
-        hash = 73 * hash + this.id;
+        hash = 29 * hash + this.id;
         return hash;
     }
 
@@ -105,7 +136,7 @@ public class ServiceSale implements Serializable {
         if (getClass() != obj.getClass()) {
             return false;
         }
-        final ServiceSale other = (ServiceSale) obj;
+        final BillPayment other = (BillPayment) obj;
         if (this.id != other.id) {
             return false;
         }
