+++ src/com/liquidbol/commons/model/Company.java
@@ -4,7 +4,7 @@
  * and open the template in the editor.
  */
 
-package com.liquidbol.model.commons;
+package com.liquidbol.commons.model;
 
 import java.util.Collection;
 import java.util.HashSet;
