+++ src/com/liquidbol/commons/model/ServiceSale.java
@@ -4,23 +4,18 @@
  * and open the template in the editor.
  */
 
-package com.liquidbol.model.commons;
+package com.liquidbol.commons.model;
 
-import com.liquidbol.db.persistence.EmployeeCrud;
-import com.liquidbol.db.persistence.PersistenceException;
 import java.io.Serializable;
 import java.sql.Date;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 /**
- * Class that represents a bill payment.
+ * Class that represents a service sale.
  * @author Allan Leon
  */
-public class BillPayment implements Serializable {
+public class ServiceSale implements Serializable {
     
     private int id;
-    private Employee employee;
     private Date payDate;
     private Double amountPaid;
     private String obs;
@@ -28,14 +23,12 @@ public class BillPayment implements Serializable {
     /**
      * Constructor method.
      * @param id
-     * @param employee
      * @param payDate
      * @param amountPaid
      * @param obs 
      */
-    public BillPayment(int id, Employee employee, Date payDate, Double amountPaid, String obs) {
+    public ServiceSale(int id, Date payDate, Double amountPaid, String obs) {
         this.id = id;
-        this.employee = employee;
         this.payDate = payDate;
         this.amountPaid = amountPaid;
         this.obs = obs;
@@ -46,13 +39,6 @@ public class BillPayment implements Serializable {
      */
     public int getId() {
         return id;
-    }
-
-    /**
-     * @return the employee
-     */
-    public Employee getEmployee() {
-        return employee;
     }
 
     /**
@@ -84,13 +70,6 @@ public class BillPayment implements Serializable {
     }
 
     /**
-     * @param employee the employee to set
-     */
-    public void setEmployee(Employee employee) {
-        this.employee = employee;
-    }
-
-    /**
      * @param payDate the payDate to set
      */
     public void setPayDate(Date payDate) {
@@ -110,21 +89,11 @@ public class BillPayment implements Serializable {
     public void setObs(String obs) {
         this.obs = obs;
     }
-    
-    public void refresh() {
-        try {
-            employee = new EmployeeCrud().refresh(employee);
-        } catch (PersistenceException ex) {
-            Logger.getLogger(BillPayment.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (ClassNotFoundException ex) {
-            Logger.getLogger(BillPayment.class.getName()).log(Level.SEVERE, null, ex);
-        }
-    }
 
     @Override
     public int hashCode() {
         int hash = 3;
-        hash = 29 * hash + this.id;
+        hash = 73 * hash + this.id;
         return hash;
     }
 
@@ -136,7 +105,7 @@ public class BillPayment implements Serializable {
         if (getClass() != obj.getClass()) {
             return false;
         }
-        final BillPayment other = (BillPayment) obj;
+        final ServiceSale other = (ServiceSale) obj;
         if (this.id != other.id) {
             return false;
         }
