+++ src/com/liquidbol/db/persistence/ItemDiscountCrud.java
@@ -6,8 +6,8 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.model.commons.Discount;
-import com.liquidbol.model.commons.Item;
+import com.liquidbol.commons.model.Discount;
+import com.liquidbol.commons.model.Item;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
