+++ src/com/liquidbol/db/persistence/RechargeableItemCrud.java
@@ -6,8 +6,8 @@
 
 package com.liquidbol.db.persistence;
 
-import com.liquidbol.model.commons.Client;
-import com.liquidbol.model.commons.RechargeableItem;
+import com.liquidbol.commons.model.Client;
+import com.liquidbol.commons.model.RechargeableItem;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
